+++ changelog.md
@@ -9,6 +9,7 @@
 * #526 Min\max amount option for which the payment method should be offered during checkout
 * #718 ShopConnector: Import option for "Published" and "Disable buy\wishlist button"
 * #702 Facebook and Twitter external authentication suitable for multi-stores
+* #704 Make primary store currency suitable for multi-stores
 
 ### Improvements
 * (Perf) Implemented static caches for URL aliases and localized properties. Increases app startup and request speed by up to 30%.